+++ Aikhsu.txt
@@ -3,3 +3,4 @@ E | false | deadline | 21/01/2002 19:00 | 21:59
 D | true | clean room clean room | 21/08/2015 19:00
 E | false | event fly high | 21/08/2002 19:00 | 13:00
 T | true | skdfjslfjksdfsfjsf
+D | false | do things | 21/02/2015 19:00
