+++ src/main/java/Commands/ListCommand.java
@@ -1,7 +1,9 @@
+package Commands;
+
 public class ListCommand extends Command{
     @Override
     public CommandResult execute() {
         tasks.listTasks();
-        return new CommandResult("Command executed successfully!");
+        return new CommandResult("Commands.Command executed successfully!");
     }
 }
