+++ src/main/java/Commands/MarkCommand.java
@@ -1,22 +1,26 @@
-public class DeleteCommand extends Command {
+package Commands;
+
+import Exceptions.AikhsuException;
+
+public class MarkCommand extends Command {
     private static final String INVALID_INDEX_MESSAGE = "Please provide a number!";
-    private static final String TASK_NOT_EXISTS_MESSAGE = "Task does not exist!";
+    private static final String TASK_NOT_EXISTS_MESSAGE = "Tasks.Task does not exist!";
     private final String indexString;
 
-    public DeleteCommand(String indexString) {
+    public MarkCommand(String indexString) {
         this.indexString = indexString.strip();
     }
 
     @Override
     public CommandResult execute() {
         try {
-            int id = Integer.parseInt(indexString) - 1;
-            this.tasks.deleteTask(id);
+            int id = Integer.parseInt(indexString);
+            this.tasks.markTask(id);
         } catch (NumberFormatException e) {
             return new CommandResult(INVALID_INDEX_MESSAGE);
         } catch (AikhsuException e) {
             return new CommandResult(TASK_NOT_EXISTS_MESSAGE);
         }
-        return new CommandResult("Task deleted successfully!");
+        return new CommandResult("Tasks.Task marked successfully!");
     }
 }
