+++ src/main/java/Exceptions/AikhsuException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class AikhsuException extends Exception{
     public AikhsuException(String message) {
         super(message);
