+++ src/main/java/Tasks/Event.java
@@ -1,8 +1,8 @@
-import java.time.LocalDate;
+package Tasks;
+
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Date;
 
 public class Event extends Task {
