+++ src/main/java/Tasks/Task.java
@@ -1,3 +1,5 @@
+package Tasks;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
